add viewModel for restaurants
